Update class SlipoContext
Replaced several methods with methods from the utils module.
Added parameter overwrite to file download method.
Added support for uploading all the files in a folder with a single method call.
Renamed the column names in several DataFrames.
Method process status can be invoked using an instance of the Process class.
Added support for using the output of any step from any workflow as input to an operation.
Instances of StepFile can be used instead of tuples for representing method arguments.
Method process file download now checks if target file name already exists or if the path is a directory.
Added methods for exporting RDF datasets to CSV or SHAPEFILE files.
Added TripleGeo parameters mappingSpec and classificationSpec to the transform operations.
Updated documentation.
Increased version to 0.1.2.